+++ Sources/Observer.swift
@@ -11,9 +11,16 @@ extension Signal {
 	/// (typically from a Signal).
 	public final class Observer {
 		public typealias Action = (Event) -> Void
+		private let _send: Action
 
 		/// An action that will be performed upon arrival of the event.
-		public let action: Action
+		@available(*, deprecated: 2.0, renamed:"send(_:)")
+		public var action: Action {
+			guard !interruptsOnDeinit && wrapped == nil else {
+				return { self._send($0) }
+			}
+			return _send
+		}
 
 		/// Whether the observer should send an `interrupted` event as it deinitializes.
 		private let interruptsOnDeinit: Bool
@@ -40,9 +47,9 @@ extension Signal {
 		) {
 			var hasDeliveredTerminalEvent = false
 
-			self.action = transform { event in
+			self._send = transform { event in
 				if !hasDeliveredTerminalEvent {
-					observer.action(event)
+					observer._send(event)
 
 					if event.isTerminating {
 						hasDeliveredTerminalEvent = true
@@ -63,7 +70,7 @@ extension Signal {
 		///   - interruptsOnDeinit: `true` if the observer should send an `interrupted`
 		///                         event as it deinitializes. `false` otherwise.
 		internal init(action: @escaping Action, interruptsOnDeinit: Bool) {
-			self.action = action
+			self._send = action
 			self.wrapped = nil
 			self.interruptsOnDeinit = interruptsOnDeinit
 		}
@@ -74,7 +81,7 @@ extension Signal {
 		/// - parameters:
 		///   - action: A closure to lift over received event.
 		public init(_ action: @escaping Action) {
-			self.action = action
+			self._send = action
 			self.wrapped = nil
 			self.interruptsOnDeinit = false
 		}
@@ -116,7 +123,7 @@ extension Signal {
 			self.init { event in
 				switch event {
 				case .value, .completed, .failed:
-					observer.action(event)
+					observer.send(event)
 				case .interrupted:
 					observer.sendCompleted()
 				}
@@ -128,8 +135,13 @@ extension Signal {
 				// Since `Signal` would ensure that only one terminal event would ever be
 				// sent for any given `Signal`, we do not need to assert any condition
 				// here.
-				action(.interrupted)
+				_send(.interrupted)
 			}
+		}
+
+		/// Puts an event into `self`.
+		public func send(_ event: Event) {
+			_send(event)
 		}
 
 		/// Puts a `value` event into `self`.
@@ -137,7 +149,7 @@ extension Signal {
 		/// - parameters:
 		///   - value: A value sent with the `value` event.
 		public func send(value: Value) {
-			action(.value(value))
+			_send(.value(value))
 		}
 
 		/// Puts a failed event into `self`.
@@ -145,17 +157,17 @@ extension Signal {
 		/// - parameters:
 		///   - error: An error object sent with failed event.
 		public func send(error: Error) {
-			action(.failed(error))
+			_send(.failed(error))
 		}
 
 		/// Puts a `completed` event into `self`.
 		public func sendCompleted() {
-			action(.completed)
+			_send(.completed)
 		}
 
 		/// Puts an `interrupted` event into `self`.
 		public func sendInterrupted() {
-			action(.interrupted)
+			_send(.interrupted)
 		}
 	}
 }